requirements.txt
websocket-client>=1.6.1
requests>=2.31.0
python-dateutil>=2.8.2
tzlocal>=5.1
websockets>=12.0
pandas>=2.0.3
colorama>=0.4.6
scikit-learn>=1.4.2
python-dotenv>=1.0.1
telethon>=1.32.0